requirements.txt
opencv-python==4.8.1.78
qrcode==7.4.2
Pillow==10.0.1
numpy==1.24.3
